Update tools to TS
